pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "ppt-pipeline"
version = "0.1.0"
description = "PDF thesis → defense PPT auto-generation pipeline"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "you@example.com"},
]
keywords = ["pdf", "pptx", "thesis", "presentation", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Text Processing",
]
dependencies = [
    "PyMuPDF>=1.23.0",
    "pdfplumber>=0.10.0",
    "python-pptx>=0.6.21",
]

[project.optional-dependencies]
llm = [
    "anthropic>=0.30.0",
    "openai>=1.0.0",
]

[project.scripts]
ppt-pipeline = "run:main"

[tool.setuptools.packages.find]
include = ["src*"]
